+++ src/core/TimeSensitive.h
@@ -23,7 +23,7 @@ public:
 	virtual ~TimeSensitive();
 	//TimeSensitive(const TimeSensitive &other);
 
-	virtual void setT(const long double& dt);	//Used to prepare the logic behind the scenes
+	virtual void setT(const float& dt);	//Used to prepare the logic behind the scenes
 	virtual void apply();						//Can used to execute an action (Ex: For the grav rule, move...)
 
 	virtual std::string to_string(const bool& spread=false, const bool& full_info=false, const unsigned char& indent=0) const;// :)
